config/__init__.py
"""
Configuration package for Dovos application.
"""

import os
from dotenv import load_dotenv

# Load environment variables from .env file
load_dotenv()

# === CONFIG ===
SECRET_KEY = os.getenv("SECRET_KEY", "your-secret-key-change-this-in-production")
SECURITY_PASSWORD_SALT = os.getenv("SECURITY_PASSWORD_SALT", "your-password-salt-change-this")

# PostgreSQL Configuration
DATABASE_URL = os.getenv("DATABASE_URL", "postgresql+psycopg://user:pass@localhost:5432/dovos_dev")

# === Authentication Configuration ===
# Whether authentication is required (can be disabled for local-only setups)
AUTH_ENABLED = os.getenv("AUTH_ENABLED", "true").lower() == "true"

# WebAuthn/Passkey Configuration
# The Relying Party (RP) name shown to users during passkey registration
WEBAUTHN_RP_NAME = os.getenv("WEBAUTHN_RP_NAME", "Scry")
# The RP ID should be the domain name (without protocol/port)
# For localhost development, use "localhost"
WEBAUTHN_RP_ID = os.getenv("WEBAUTHN_RP_ID", "localhost")
# The origin URL for WebAuthn (must match the URL users access the app from)
WEBAUTHN_ORIGIN = os.getenv("WEBAUTHN_ORIGIN", "http://localhost:5001")
PGAPPNAME = os.getenv("PGAPPNAME", "dovos-api")
EMBEDDING_MODEL = os.getenv("EMBEDDING_MODEL", "sentence-transformers/all-MiniLM-L6-v2")
EMBEDDING_DIM = int(os.getenv("EMBEDDING_DIM", "384"))

# RAG Context Configuration
RAG_DEFAULT_WINDOW_SIZE = int(os.getenv("RAG_WINDOW_SIZE", "3"))
RAG_MAX_WINDOW_SIZE = int(os.getenv("RAG_MAX_WINDOW_SIZE", "10"))
RAG_ADAPTIVE_WINDOWING = os.getenv("RAG_ADAPTIVE_WINDOWING", "true").lower() == "true"
RAG_DEDUPLICATE_MESSAGES = os.getenv("RAG_DEDUPLICATE_MESSAGES", "true").lower() == "true"
RAG_DEFAULT_TOP_K_WINDOWS = int(os.getenv("RAG_DEFAULT_TOP_K_WINDOWS", "8"))
RAG_DEFAULT_MAX_TOKENS = int(os.getenv("RAG_DEFAULT_MAX_TOKENS", "0"))  # 0 = no limit
RAG_PROXIMITY_DECAY_LAMBDA = float(os.getenv("RAG_PROXIMITY_DECAY_LAMBDA", "0.3"))
RAG_APPLY_RECENCY_BONUS = os.getenv("RAG_APPLY_RECENCY_BONUS", "false").lower() == "true"

# Version
def get_version():
    """Read version from VERSION file."""
    try:
        version_file = os.path.join(os.path.dirname(os.path.dirname(__file__)), "VERSION")
        with open(version_file, 'r') as f:
            return f.read().strip()
    except Exception:
        return "unknown"

VERSION = get_version()

# Import synonym functionality
from .synonyms import SEARCH_SYNONYMS, get_synonyms, add_synonym_mapping

__all__ = [
    'SECRET_KEY',
    'SECURITY_PASSWORD_SALT',
    'DATABASE_URL',
    'PGAPPNAME',
    'EMBEDDING_MODEL',
    'EMBEDDING_DIM',
    'VERSION',
    'get_version',
    'SEARCH_SYNONYMS',
    'get_synonyms',
    'add_synonym_mapping',
    'RAG_DEFAULT_WINDOW_SIZE',
    'RAG_MAX_WINDOW_SIZE',
    'RAG_ADAPTIVE_WINDOWING',
    'RAG_DEDUPLICATE_MESSAGES',
    'RAG_DEFAULT_TOP_K_WINDOWS',
    'RAG_DEFAULT_MAX_TOKENS',
    'RAG_PROXIMITY_DECAY_LAMBDA',
    'RAG_APPLY_RECENCY_BONUS',
    'AUTH_ENABLED',
    'WEBAUTHN_RP_NAME',
    'WEBAUTHN_RP_ID',
    'WEBAUTHN_ORIGIN',
]
