requirements.txt
flask
flask-cors
flask-wtf
wtforms

# Authentication with passkey support
flask-security-too[mfa]
bcrypt
sentence-transformers
markdown
python-docx
nltk
requests
uuid
# API Contract validation
pydantic>=2.0

# PostgreSQL dependencies
sqlalchemy>=2.0
psycopg[binary]>=3.0
alembic>=1.12
python-dotenv>=1.0
pgvector

# Testing dependencies
pytest>=7.0
pytest-flask
pytest-json-report
pytest-cov
responses
